refactor: Project File Structure Simplification
Removed user directory from project hierarchy to enhance organization and
simplicity.
