+++ lua/plugins/flash.lua
